requirements.txt
cffi==1.15.0
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
Flask==2.0.3
gunicorn==20.1.0
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
pycparser==2.21
PyJWT==2.3.0
Werkzeug==2.0.3
